+++ ui-test/tests/Selects.jsx
@@ -1,112 +1,120 @@
 import React from 'react';
+import _ from 'lodash';
+import {
+    Card,
+    CardTitle,
+    CardContent,
+    SelectBox,
+    AutoCompleteBox,
+    Checkbox,
+} from '../../index';
 
-import {Card, CardTitle, CardContent, SelectBox} from '../../index';
+const selectOptions = [
+    {
+        label: 'Label 1',
+        value: 'http://example.com/one',
+        description: 'This is one of the options with a description',
+    },
+    {
+        label: 'label two',
+        value: 'http://example.com/two',
+    },
+    {
+        label: 'label three',
+        value: 'http://example.com/three',
+        description: 'This is one of the options with another description',
+    },
+    {
+        label: 'http://example.com/four',
+        value: 'http://example.com/four',
+        description: 'The fourth option has the same data for label and value',
+    },
+];
 
-const TestInputs = React.createClass({
+const TestSelects = React.createClass({
     getInitialState() {
         return {
-            selectBox1: {label: 'labelz', value: 'valuez'},
-            selectBox2: 8,
-            selectBox3: '',
-            selectBox4: '',
+            selectSettings: {
+                searchable: true,
+                optionsOnTop: false,
+                clearable: false,
+                multi: true,
+                creatable: false,
+            },
+            selectBox1: _.head(selectOptions),
+            selectBox2: [],
+            autoCompleteBox: null,
         };
     },
-    handleChange(value, key) {
+    handleSelectChange(value, key) {
         console.log('SelectBox onChange: ', value);
         this.setState({[key]: value});
     },
+    handleChange({value, name}) {
+        const newState = this.state;
+        console.warn(newState, name);
+        _.set(newState, name, value);
+        this.setState(newState);
+    },
     render() {
+        const {selectSettings} = this.state;
+
         return (
             <Card>
                 <CardTitle documentLevel={'h4'}>Test Selects</CardTitle>
                 <CardContent>
-                    <h5>SelectBox with objects</h5>
+                    <h5>Select Boxes</h5>
                     <SelectBox
-                        placeholder={'Value deleted'}
-                        options={[
-                            {
-                                label: 'labelz',
-                                value: 'valuez',
-                            },
-                            {
-                                label: 'label1',
-                                value: 'value1',
-                            },
-                            {
-                                label: 'label2',
-                                value: 'value2',
-                            },
-                        ]}
+                        placeholder={'Sync select box'}
+                        options={selectOptions}
                         name="selectBox1"
                         value={this.state.selectBox1}
-                        onChange={this.handleChange}
-                        autofocus
+                        onChange={this.handleSelectChange}
+                        {...selectSettings}
                     />
-                    <h5>SelectBox with mixed strings and numbers</h5>
                     <SelectBox
-                        placeholder={'No Value'}
-                        options={['label1', 3, 8]}
+                        placeholder={'Async select box'}
                         name="selectBox2"
                         value={this.state.selectBox2}
-                        onChange={this.handleChange}
-                    />
-                    <SelectBox
-                        placeholder={'Open option to top'}
-                        options={['label1', 3, 8]}
-                        name="selectBox2"
-                        value={this.state.selectBox2}
-                        onChange={this.handleChange}
-                        optionsOnTop
-                    />
-                    <h5>
-                        Multiple selections with option to create new entries
-                    </h5>
-                    <SelectBox
-                        placeholder={'Multiselect with create'}
-                        options={[
-                            'label1',
-                            3,
-                            'Andelsselskab med begrænset ansvar, Denmark',
-                            'Akcionersko Društvo / Акционерско друштво, Republic of Macedonia',
-                            8,
-                        ]}
-                        name="selectBox3"
-                        value={this.state.selectBox3}
-                        onChange={this.handleChange}
-                        multi
-                        clearable={false}
-                        creatable
-                        promptTextCreator={newLabel =>
-                            `New freaking Stuff called: ${newLabel}`
-                        }
-                    />
-                    <h5>Async Selectbox</h5>
-                    <SelectBox
-                        placeholder={'Multiselect with async'}
-                        name="selectBox4"
-                        value={this.state.selectBox4}
-                        onChange={this.handleChange}
+                        onChange={this.handleSelectChange}
+                        async
                         loadOptions={function(input, callback) {
+                            console.warn('loadOptions');
                             setTimeout(() => {
                                 callback(null, {
-                                    options: [
-                                        {value: 'one', label: 'One'},
-                                        {value: 'two', label: 'Two'},
-                                    ],
+                                    options: selectOptions,
                                     // CAREFUL! Only set this to true when there are no more options,
                                     // or more specific queries will not be sent to the server.
                                     complete: true,
                                 });
-                            }, 500);
+                            }, 250);
                         }}
-                        multi
-                        async
+                        {...selectSettings}
                     />
                     <hr />
+                    Settings:
+                    {_.map(selectSettings, (value, key) => (
+                        <Checkbox
+                            key={key}
+                            name={`selectSettings.${key}`}
+                            checked={value}
+                            onChange={this.handleChange}>
+                            {key}
+                        </Checkbox>
+                    ))}
+                    <hr />
+                    <h5>Auto Complete Box</h5>
+                    <AutoCompleteBox
+                        placeholder={'Sync select box'}
+                        options={selectOptions}
+                        name="autoCompleteBox"
+                        value={this.state.autoCompleteBox}
+                        onChange={this.handleSelectChange}
+                    />
                 </CardContent>
             </Card>
         );
     },
 });
 
-export default TestInputs;
+export default TestSelects;